extract the scheduler from handler package
Introduces a new Scheduler interface that holds ScheduleTask
