+++ eremetic.go
@@ -6,12 +6,13 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/alde/eremetic/database"
-	"github.com/alde/eremetic/handler"
-	"github.com/alde/eremetic/routes"
 	log "github.com/dmuth/google-go-log4go"
 	"github.com/kardianos/osext"
 	"github.com/spf13/viper"
+
+	"github.com/alde/eremetic/database"
+	"github.com/alde/eremetic/routes"
+	"github.com/alde/eremetic/scheduler"
 )
 
 func readConfig() {
@@ -56,7 +57,7 @@ func main() {
 
 	router := routes.Create()
 	log.Infof("listening to %s", bind)
-	go handler.Run()
+	go scheduler.Run()
 	err := http.ListenAndServe(bind, router)
 	if err != nil {
 		log.Error(err.Error())
