+++ scheduler/driver_test.go
@@ -1,4 +1,4 @@
-package handler
+package scheduler
 
 import (
 	"testing"
