+++ types/scheduler.go
@@ -1,0 +1,5 @@
+package types
+
+type Scheduler interface {
+	ScheduleTask(request Request) (string, error)
+}
